+++ my_solved/lib/pages/main_tab_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../veiws/main_tab_view.dart';
+import '../views/main_tab_view.dart';
 import '../view_models/main_tab_view_model.dart';
 
 class MainTabPage extends StatelessWidget {
