+++ my_solved/lib/pages/problem_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:my_solved/veiws/problem_view.dart';
+import 'package:my_solved/views/problem_view.dart';
 import 'package:my_solved/view_models/problem_view_model.dart';
 
 class ProblemPage extends StatelessWidget {
