+++ my_solved/lib/pages/search_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../veiws/search_view.dart';
+import '../views/search_view.dart';
 import '../view_models/search_view_model.dart';
 
 class SearchPage extends StatelessWidget {
