+++ my_solved/lib/providers/search/problem_api.dart
@@ -2,14 +2,16 @@ import 'dart:async';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 
-import 'package:my_solved/model/search/problem.dart';
+import 'package:my_solved/models/search/problem.dart';
 
 Future<SearchProblem> searchProblem() async {
-  final response = await http.get(Uri.parse('https://solved.ac/api/v3/search/suggestion?query=rsa'));
+  final response = await http
+      .get(Uri.parse('https://solved.ac/api/v3/search/suggestion?query=rsa'));
   final statusCode = response.statusCode;
 
-  if (statusCode == 200){
-    SearchProblem searchedProblems = SearchProblem.fromJson(jsonDecode(response.body));
+  if (statusCode == 200) {
+    SearchProblem searchedProblems =
+        SearchProblem.fromJson(jsonDecode(response.body));
     return searchedProblems;
   } else {
     throw Exception('Failed to load');
