+++ my_solved/lib/providers/search/suggestion_api.dart
@@ -2,10 +2,11 @@ import 'dart:async';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 
-import 'package:my_solved/model/search/suggestion.dart';
+import 'package:my_solved/models/search/suggestion.dart';
 
 Future<SearchSuggestion> searchSuggestion(String query) async {
-  final response = await http.get(Uri.parse("https://solved.ac/api/v3/search/suggestion?query=$query"));
+  final response = await http.get(
+      Uri.parse("https://solved.ac/api/v3/search/suggestion?query=$query"));
   final statusCode = response.statusCode;
 
   if (statusCode == 200) {
