+++ my_solved/lib/view_models/problem_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:my_solved/provider/acmicpc/problem_api.dart';
+import 'package:my_solved/providers/acmicpc/problem_api.dart';
 import 'package:html/dom.dart' as dom;
 
 class ProblemViewModel with ChangeNotifier {
